requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.5.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0
pyyaml>=5.4.0
albumentations>=1.0.0